third_party/WebKit/LayoutTests/fast/canvas-api/canvas-createImageBitmap-invalid-args-in-workers-expected.txt

[Worker] Test createImageBitmap with invalid arguments in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ./resources/canvas-createImageBitmap-invalid-args-in-workers.js
PASS [Worker] Rejected as expected: null
PASS [Worker] reason instanceof Error is true
[Worker] TypeError: Failed to execute 'createImageBitmap' on 'WorkerGlobalScope': The provided value is not of type '(HTMLImageElement or SVGImageElement or HTMLVideoElement or HTMLCanvasElement or Blob or ImageData or ImageBitmap or OffscreenCanvas)'
PASS [Worker] Rejected as expected: invalid area
PASS [Worker] reason instanceof Error is true
[Worker] RangeError: The crop rect height is 0.
PASS [Worker] Rejected as expected: invalid area
PASS [Worker] reason instanceof Error is true
[Worker] RangeError: The crop rect width is 0.
PASS successfullyParsed is true

TEST COMPLETE
